プッシュ

+++ src/main/java/com/example/outputmanager/domain/Favorite.java
@@ -1,12 +1,13 @@
 package com.example.outputmanager.domain;
 
-import lombok.Data;
 import java.time.LocalDateTime;
+
+import lombok.Data;
 
 @Data
 public class Favorite {
-    private Integer id;
-    private Integer userId;
-    private Integer outputId;
-    private LocalDateTime createdAt;
+    private Integer id;          // favorites.id（INTのままでOK）
+    private Integer userId;      // favorites.user_id
+    private Long outputId;       // ★ favorites.output_id → Output.id(BIGINT) に合わせて Long
+    private LocalDateTime createdAt; // favorites.created_at
 }
